Move specialization_graph definition in traits::types.
