+++ src/librustc/traits/types/mod.rs
@@ -4,6 +4,7 @@
 
 pub mod query;
 pub mod select;
+pub mod specialization_graph;
 mod structural_impls;
 
 use crate::mir::interpret::ErrorHandled;
